requirement.txt
streamlit
mysql-connector-python
python-dotenv
pandas
openpyxl
groq
Pillow
spacy